Only show zheight label when printer has been homed
this is to prevent misleading the user - the zheight isn't tracked by
photonic3d (and technically cannot be, as it relies on the buildplate
not being manually handled or adjusted at any point to say, remove a
print.)
so hide the label until the printer has been homed. at that point, show
information about the z-height.