Printflow.js now tries to query Wifi information from Photonic
As the build doesn't have this integrated (I'm waiting on a code review
of my changes) wifi symbol will show red, even if it does have a wifi
connection. As the wifi symbol was previously generating random
information anyway, changing this ready for week 3 deliverables is
non-destructive.

Also, work to support a better version of the printdialogue page.