Improvement for start.sh not auto-updating on boot
This now tests for a connection to github. It will wait for up to 20 seconds [tunable on line 105] for a network connection.

If it's found, it will do all the logic for online updating. 
If it isn't found, it'll push ahead with launching Photonic.

This will introduce some extra time into the raspberry pi boot sequence, but as the DHCP information isn't available 99% of the times when this script is reaching a stage that needs it, it's better to do it this way and enable online updating.